Extract rate-limited interaction handling into shared helper
Reduce duplication between app_mention.js and message.js by extracting the rate-limited interaction path into a reusable handleRateLimitedInteraction() function. This centralizes telemetry recording, error handling, and user messaging, lowering maintenance drift risk when these fields evolve.
